Implement reportGoalResult.  Fix memory management (autoreleasepool) of attemptGoal callbacks
